Refactor: 업로드 로직 수정

See merge request s11-webmobile1-sub2/S11P12A505!67

+++ backend/src/main/java/com/ssafy/a505/domain/service/MemberServiceImpl.java
@@ -56,8 +56,11 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public boolean removeUser(long memberId) {
-        memberRepository.deleteById(memberId);
-        return true;
+        if (memberRepository.existsById(memberId)) {
+            memberRepository.deleteById(memberId);
+            return true;
+        }
+        return false;
     }
 
     @Override
